refactor(snackbar): global providerからsnack関数を取得する
