+++ packages/wiz-ui/src/hooks/use-global-provider.ts
@@ -1,38 +1,7 @@
-import { ref, readonly, InjectionKey, inject } from "vue";
+import { InjectionKey, inject } from "vue";
 
-import { SnackbarOption } from "@/components/base/snackbar/types";
-
-const useMenu = () => {
-  const isMenuOpen = ref(true);
-  const setIsMenuOpen = (value: boolean) => {
-    isMenuOpen.value = value;
-  };
-  return {
-    isMenuOpen: readonly(isMenuOpen),
-    setIsMenuOpen,
-  };
-};
-
-const useSnackbar = () => {
-  const options = ref<SnackbarOption[]>([]);
-  const snack = (message: string) => {
-    const created = new Date().toISOString();
-    const deleteSnackbar = () => {
-      options.value = options.value.filter(
-        (options) => options.created !== created
-      );
-    };
-    options.value.push({
-      message,
-      created,
-      delete: deleteSnackbar,
-    });
-  };
-  return {
-    snackbarOptions: readonly(options),
-    snack,
-  };
-};
+import { useMenu } from "./use-menu";
+import { useSnackbar } from "./use-snackbar";
 
 export const useGlobalProvider = () => {
   return {
