+++ packages/wiz-ui/src/hooks/use-menu.ts
@@ -1,0 +1,12 @@
+import { ref, readonly } from "vue";
+
+export const useMenu = () => {
+  const isMenuOpen = ref(true);
+  const setIsMenuOpen = (value: boolean) => {
+    isMenuOpen.value = value;
+  };
+  return {
+    isMenuOpen: readonly(isMenuOpen),
+    setIsMenuOpen,
+  };
+};
